requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
markitdown[all]>=0.1.0
flask>=3.0.0
